code upgrade on fuction 'periodogram'
